requirements.txt
cvxpy==1.1
Cython==0.29.23
imageio==2.15.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
motmetrics==1.2.5
numba==0.53.1
numpy==1.18.1
opencv-python==4.5.*
pandas==1.1.5
Pillow==8.4.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
Shapely==1.7.1
silence-tensorflow==1.1.1
sklearn==0.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.2
tifffile==2020.9.3
tqdm==4.54.0
rtree
cylp
